+++ indico/modules/events/layout/forms.py
@@ -11,13 +11,13 @@ from wtforms.validators import DataRequired, Optional, ValidationError
 
 from indico.core.config import config
 from indico.modules.events.layout import theme_settings
-from indico.modules.events.layout.models.menu import MenuEntryAccess
+from indico.modules.events.layout.models.menu import MenuEntry
 from indico.modules.events.layout.util import get_css_file_data, get_logo_data, get_plugin_conference_themes
 from indico.modules.users import NameFormat
 from indico.util.i18n import _, orig_string
 from indico.web.forms.base import IndicoForm
-from indico.web.forms.fields import EditableFileField, FileField, IndicoEnumSelectField
-from indico.web.forms.fields.enums import IndicoEnumRadioField
+from indico.web.forms.fields import EditableFileField, FileField, IndicoEnumSelectField, IndicoProtectionField
+from indico.web.forms.fields.principals import PrincipalListField
 from indico.web.forms.validators import HiddenUnless, UsedIf
 from indico.web.forms.widgets import CKEditorWidget, ColorPickerWidget, SwitchWidget
 
@@ -163,8 +163,25 @@ class MenuUserEntryFormBase(IndicoForm):
     title = StringField(_('Title'), [DataRequired()])
     is_enabled = BooleanField(_('Show'), widget=SwitchWidget())
     new_tab = BooleanField(_('Open in a new tab'), widget=SwitchWidget())
-    access = IndicoEnumRadioField(_('Visibility'), enum=MenuEntryAccess,
-                                  description=_('Specify who can see this menu item.'))
+    acl = PrincipalListField(
+        _('Visibility'),
+        event=lambda form: form.event,
+        allow_groups=True,
+        allow_event_roles=True,
+        allow_category_roles=True,
+        allow_registration_forms=True,
+    )
+    speaker_allowed = BooleanField(
+        _('Speaker Visibility'),
+        widget=SwitchWidget(),
+        description=_('In addition to the visibility settings, speakers will be allowed access to this entry'),
+    )
+    protection_mode = IndicoProtectionField(_('Protection mode'), protected_object=lambda form: form.protected_object)
+
+    def __init__(self, *args, event, **kwargs):
+        self.event = event
+        self.protected_object = kwargs.get('entry', MenuEntry(event=event))
+        super().__init__(*args, **kwargs)
 
 
 class MenuLinkForm(MenuUserEntryFormBase):