+++ CHANGES.rst
@@ -215,6 +215,8 @@ Improvements
   (:pr:`5784`)
 - Allow admins to change the password of local accounts (:pr:`5789`, thanks
   :user:`omegak`)
+- Menu entries now have full ACLs. This can be used to created pages that are restricted
+  to certain users, groups, registrations, or roles. (:pr:`5670`, thanks :user:`kewisch`)
 
 Bugfixes
 ^^^^^^^^